Removed all but one class - ini
- broken down main class into a bunch of probably dead functions
- discovered the next step in this overhaul

At the moment there are no values being added to the gui where as they should. So I'm presuming there is something wrong with the ini class, which isn't the end of the world.
The good news however is that gui does indeed appear! Hopefully if I recover the ini class and update all references to it then I'll be one step closer to completion.